+++ CHANGELOG.md
@@ -8,6 +8,8 @@ the release.
 ## Unreleased
 
 - API improvements - use C# native classes to measure time.
+- OpenCensus.Collectors.AspNetCore: Allow to supply custom sampler based on request properties using custom code. For instance filter out telemetry from specific path.
+- OpenCensus.Collectors.Dependencies: Allow to supply custom sampler based on request properties using custom code. By default, filter out calls to Zipkin REST endpoint from the exporter.
 
 ## 0.1.0-alpha-42253